@@ -5,46 +5,44 @@ import 'package:voyager/voyager.dart';
 import 'package:voyager_bloc/voyager_bloc.dart';
 
 class ParentBloc extends Bloc {
-  @override
-  get initialState => null;
+
+  ParentBloc() : super(null);
 
   @override
   Stream mapEventToState(event) {
-    return null;
+    return event;
   }
 }
 
 class StrangerBloc extends Bloc {
-  @override
-  get initialState => null;
+
+  StrangerBloc() : super(null);
 
   @override
   Stream mapEventToState(event) {
-    return null;
+    return event;
   }
 }
 
 class ChildBloc extends ParentBloc {
-  @override
-  get initialState => null;
+
+  ChildBloc() : super();
+
 
   @override
   Stream mapEventToState(event) {
-    return null;
+    return event;
   }
 }
 
 class CounterBloc extends Bloc {
-  final int _initialValue;
+  final int? initialValue;
 
-  CounterBloc(this._initialValue);
-
-  @override
-  get initialState => _initialValue;
+  CounterBloc({this.initialValue}) : super(null);
 
   @override
   Stream mapEventToState(event) {
-    return null;
+    return event;
   }
 }
 
@@ -53,7 +51,7 @@ void main() {
     final builder = BlocsPluginBuilder()
         .addBaseBloc<ParentBloc>((context, config, repository) => ParentBloc())
         .addBaseBloc<CounterBloc>((context, config, repository) =>
-            CounterBloc(int.parse(config.toString())))
+            CounterBloc(initialValue: int.parse(config.toString())))
         .addBloc<ChildBloc, ParentBloc>(
             (context, config, repository) => ChildBloc())
         .addBaseBloc<StrangerBloc>(
@@ -61,7 +59,7 @@ void main() {
 
     final blocPlugin = builder.build();
 
-    final output = Voyager(config: {});
+    final output = Voyager(config: {},path: "",pathParams: {});
     blocPlugin.outputFor(
         null,
         [
@@ -99,7 +97,7 @@ void main() {
         () => BlocsPluginBuilder().addBaseBloc(null),
         throwsA(allOf(
             isArgumentError,
-            predicate((e) =>
+            predicate((dynamic e) =>
                 e.message ==
                 'BlocType must be a subclass of BlocParentType'))));
 
@@ -107,7 +105,7 @@ void main() {
         () => BlocsPluginBuilder().addBloc<ChildBloc, Bloc>(null),
         throwsA(allOf(
             isArgumentError,
-            predicate((e) =>
+            predicate((dynamic e) =>
                 e.message == 'BlocParentType must be a subclass of Bloc'))));
   });
 
@@ -117,7 +115,7 @@ void main() {
 
     final blocPlugin = builder.build();
 
-    final output = Voyager(config: {});
+    final output = Voyager(config: {},path: "",pathParams: {});
     expect(
         () => blocPlugin.outputFor(
             null,
@@ -127,7 +125,7 @@ void main() {
             output),
         throwsA(allOf(
             isArgumentError,
-            predicate((e) =>
+            predicate((dynamic e) =>
                 e.message == 'Too many @ sings in the key of the Bloc'))));
   });
 
@@ -136,7 +134,7 @@ void main() {
 
     final blocPlugin = builder.build();
 
-    final output = Voyager(config: {});
+    final output = Voyager(config: {},path: "",pathParams: {});
     expect(
         () => blocPlugin.outputFor(
             null,
@@ -145,14 +143,14 @@ void main() {
             ],
             output),
         throwsA(allOf(isUnimplementedError,
-            predicate((e) => e.message == 'No bloc builder for ParentBloc'))));
+            predicate((dynamic e) => e.message == 'No bloc builder for ParentBloc'))));
   });
 
   test('bloc additive builders', () {
     final builderA = BlocsPluginBuilder()
         .addBaseBloc<ParentBloc>((context, config, repository) => ParentBloc())
         .addBaseBloc<CounterBloc>((context, config, repository) =>
-            CounterBloc(int.parse(config.toString())));
+            CounterBloc(initialValue: int.parse(config.toString())));
 
     final builderB = BlocsPluginBuilder()
         .addBloc<ChildBloc, ParentBloc>(
@@ -165,7 +163,7 @@ void main() {
 
     final blocPlugin = builder.build();
 
-    final output = Voyager(config: {});
+    final output = Voyager(config: {},path: "",pathParams: {});
     blocPlugin.outputFor(
         null,
         [
