@@ -1,12 +1,12 @@
-/// Generated file, DO NOT EDIT
-import 'package:flutter/widgets.dart';
-import 'package:voyager/voyager.dart';
-import 'package:voyager_bloc/voyager_bloc.dart';
+// /// Generated file, DO NOT EDIT
+// import 'package:flutter/widgets.dart';
+// import 'package:voyager/voyager.dart';
+// import 'package:voyager_bloc/voyager_bloc.dart';
 
-const String pathCounter = "/counter";
-const String typeCounter = "counter";
+// const String pathCounter = "/counter";
+// const String typeCounter = "counter";
 
-extension VoyagerData on Voyager {
-  BlocRepository get blocs => this["blocs"];
-  String get title => this["title"];
-}
+// extension VoyagerData on Voyager {
+//   BlocRepository get blocs => this["blocs"];
+//   String get title => this["title"];
+// }
