@@ -10,7 +10,8 @@ import 'package:voyager_bloc/voyager_bloc.dart';
 
 import './gen/voyager_gen.dart';
 
-class SimpleBlocDelegate extends BlocDelegate {
+class SimpleBlocDelegate extends BlocObserver {
+
   @override
   void onEvent(Bloc bloc, Object event) {
     super.onEvent(bloc, event);
@@ -23,14 +24,14 @@ class SimpleBlocDelegate extends BlocDelegate {
     print(transition);
   }
 
-  @override
-  void onError(Bloc bloc, Object error, StackTrace stacktrace) {
-    super.onError(bloc, error, stacktrace);
-    print(error);
-  }
+  // @override
+  // void onError(Bloc bloc, Object error, StackTrace stacktrace) {
+  //   super.onError(bloc, error, stacktrace);
+  //   print(error);
+  // }
 }
 
-final paths = loadPathsFromString('''
+final paths = loadPathsFromYamlString('''
 '/counter' :
   type: counter
   title: Counter
@@ -41,7 +42,7 @@ final paths = loadPathsFromString('''
 ''');
 
 final plugins = [
-  WidgetPluginBuilder().add<CounterPage>((context) => CounterPage()).build(),
+  WidgetPluginBuilder().add("CounterPage",(context) => CounterPage()).build(),
   BlocsPluginBuilder()
       .addBaseBloc<CounterBloc>(
           (context, config, repo) => CounterBloc.fromConfig(config))
@@ -51,7 +52,8 @@ final plugins = [
 ];
 
 void main() async {
-  BlocSupervisor.delegate = SimpleBlocDelegate();
+  // 5.0.0 BREAKING: Remove BlocSupervisor and rename BlocDelegate to BlocObserver.
+  // BlocSupervisor.delegate = SimpleBlocDelegate();
   runApp(Provider.value(
     value: await loadRouter(paths, plugins),
     child: App(),
@@ -63,7 +65,7 @@ class App extends StatelessWidget {
   Widget build(BuildContext context) {
     return MaterialApp(
       title: 'Flutter Demo',
-      home: VoyagerWidget(path: pathCounter),
+      home: Text("")//VoyagerWidget(path: pathCounter),
     );
   }
 }
@@ -72,80 +74,80 @@ class CounterPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final voyager = context.voyager;
+    // TODO if needed: check why voyager.blocs is not available
     // ignore: close_sinks
-    final counterBloc = voyager.blocs.find<CounterBloc>();
-    // ignore: close_sinks
-    final themeBloc = voyager.blocs.find<ThemeBloc>();
-
-    return BlocBuilder<ThemeBloc, ThemeData>(
-      bloc: themeBloc,
-      builder: (context, data) => Theme(
-        data: data,
-        child: Scaffold(
-          appBar: AppBar(title: Text(voyager.title)),
-          body: BlocBuilder<CounterBloc, int>(
-            bloc: counterBloc,
-            builder: (context, count) {
-              return Center(
-                child: Text(
-                  '$count',
-                  style: TextStyle(fontSize: 24.0),
-                ),
-              );
-            },
-          ),
-          floatingActionButton: Column(
-            crossAxisAlignment: CrossAxisAlignment.end,
-            mainAxisAlignment: MainAxisAlignment.end,
-            children: <Widget>[
-              Padding(
-                padding: EdgeInsets.symmetric(vertical: 5.0),
-                child: FloatingActionButton(
-                  child: Icon(Icons.add),
-                  onPressed: () {
-                    counterBloc.add(CounterEvent.increment);
-                  },
-                ),
-              ),
-              Padding(
-                padding: EdgeInsets.symmetric(vertical: 5.0),
-                child: FloatingActionButton(
-                  child: Icon(Icons.remove),
-                  onPressed: () {
-                    counterBloc.add(CounterEvent.decrement);
-                  },
-                ),
-              ),
-              Padding(
-                padding: EdgeInsets.symmetric(vertical: 5.0),
-                child: FloatingActionButton(
-                  child: Icon(Icons.update),
-                  onPressed: () {
-                    themeBloc.add(ThemeEvent.toggle);
-                  },
-                ),
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
+    // final counterBloc = voyager.blocs.find<CounterBloc>();
+    // // ignore: close_sinks
+    // final themeBloc = voyager.blocs.find<ThemeBloc>();
+    return Scaffold();
+  //   return BlocBuilder<ThemeBloc, ThemeData>(
+  //     bloc: themeBloc,
+  //     builder: (context, data) => Theme(
+  //       data: data,
+  //       child: Scaffold(
+  //         appBar: AppBar(title: Text(voyager.title)),
+  //         body: BlocBuilder<CounterBloc, int>(
+  //           bloc: counterBloc,
+  //           builder: (context, count) {
+  //             return Center(
+  //               child: Text(
+  //                 '$count',
+  //                 style: TextStyle(fontSize: 24.0),
+  //               ),
+  //             );
+  //           },
+  //         ),
+  //         floatingActionButton: Column(
+  //           crossAxisAlignment: CrossAxisAlignment.end,
+  //           mainAxisAlignment: MainAxisAlignment.end,
+  //           children: <Widget>[
+  //             Padding(
+  //               padding: EdgeInsets.symmetric(vertical: 5.0),
+  //               child: FloatingActionButton(
+  //                 child: Icon(Icons.add),
+  //                 onPressed: () {
+  //                   counterBloc.add(CounterEvent.increment);
+  //                 },
+  //               ),
+  //             ),
+  //             Padding(
+  //               padding: EdgeInsets.symmetric(vertical: 5.0),
+  //               child: FloatingActionButton(
+  //                 child: Icon(Icons.remove),
+  //                 onPressed: () {
+  //                   counterBloc.add(CounterEvent.decrement);
+  //                 },
+  //               ),
+  //             ),
+  //             Padding(
+  //               padding: EdgeInsets.symmetric(vertical: 5.0),
+  //               child: FloatingActionButton(
+  //                 child: Icon(Icons.update),
+  //                 onPressed: () {
+  //                   themeBloc.add(ThemeEvent.toggle);
+  //                 },
+  //               ),
+  //             ),
+  //           ],
+  //         ),
+  //       ),
+  //     ),
+  //   );
   }
 }
 
 enum CounterEvent { increment, decrement }
 
 class CounterBloc extends Bloc<CounterEvent, int> {
-  CounterBloc(this._initalState) : super();
+  CounterBloc(this.initalState) : super((){
+    return 0;
+  }());
 
   factory CounterBloc.fromConfig(dynamic config) {
     return CounterBloc(int.parse(config.toString()));
   }
 
-  final int _initalState;
-
-  @override
-  int get initialState => _initalState;
+  final int initalState;
 
   @override
   Stream<int> mapEventToState(CounterEvent event) async* {
@@ -163,7 +165,7 @@ class CounterBloc extends Bloc<CounterEvent, int> {
 enum ThemeEvent { toggle }
 
 class ThemeBloc extends Bloc<ThemeEvent, ThemeData> {
-  ThemeBloc(this._intialState) : super();
+  ThemeBloc(this.intialState) : super(null);
 
   factory ThemeBloc.fromConfig(dynamic config) {
     ThemeData data;
@@ -178,10 +180,7 @@ class ThemeBloc extends Bloc<ThemeEvent, ThemeData> {
     return ThemeBloc(data);
   }
 
-  final ThemeData _intialState;
-
-  @override
-  ThemeData get initialState => _intialState;
+  final ThemeData intialState;
 
   @override
   Stream<ThemeData> mapEventToState(ThemeEvent event) async* {
